Rework base/basic types
